RRSPH v3.1.4 - RRSPH: common computing modules between RRSPH_CL and RRSPH_OMP - RRSPH: enhance accuracy with double-precision calculations - RRSPH: 3D visualization tool

+++ sources/RRSPH_PARAMS/include/RRTypes.h
@@ -1,6 +1,6 @@
 #pragma once
 
-typedef float rr_float;
+typedef double rr_float;
 typedef unsigned long rr_uint;
 typedef long rr_int;
 typedef long rr_iter;
